rename: fetchUserInfo 경로 변경, 그로 인한 import fetchUserInfo 경로 변경 (#44)
